TABLE ASSIGNMENT
Complete these actions:

Create a new document and name it HTML_Table.html. Save this in your Basic_HTML_and_CSS folder.

In the HTML_Table.html file, utilize the <table> tag to create your own table. The rows and columns can be of your choosing.

Save the file and successfully run it in the browser.